refactor: split widgets into separate files
